PHPStan level 9

+++ src/mongo/delegates/Tables.php
@@ -11,6 +11,7 @@ use MongoDB\Collection;
 use MongoDB\Driver\CursorInterface;
 use MongoDB\Driver\Exception\BulkWriteException;
 use MongoDB\Driver\ReadPreference;
+use MongoDB\Model\BSONDocument;
 use Tripod\Config;
 use Tripod\Exceptions\LabellerException;
 use Tripod\ITripodStat;
@@ -108,8 +109,6 @@ class Tables extends CompositeBase
 
     /**
      * Returns an array of table rows that are impacted by the changes.
-     *
-     * @return mixed[]
      */
     public function findImpactedComposites(array $resourcesAndPredicates, string $contextAlias): array
     {
@@ -210,10 +209,13 @@ class Tables extends CompositeBase
      * instances, or plain arrays depending on the documentType option); pass
      * 'returnCursor' => true to receive the underlying cursor instead.
      *
-     * @param array<string, mixed> $options Table query options
-     * @param array<string, mixed> $filter
+     * @param array $options Table query options
      *
-     * @return ($options is array{returnCursor: true} ? array{head: array{count: int, offset: int|null, limit: int|null}, results: CursorInterface<int, array<string, mixed>|\ArrayObject<string, mixed>>} : array{head: array{count: int, offset: int|null, limit: int|null}, results: list<array<string, mixed>|\ArrayObject<string, mixed>>})
+     * @return (
+     *  $options is array{returnCursor: true}
+     *      ? array{head: array{count: int, offset: int|null, limit: int|null}, results: CursorInterface<int, array|BSONDocument>}
+     *      : array{head: array{count: int, offset: int|null, limit: int|null}, results: list<array|BSONDocument>}
+     * )
      */
     public function getTableRows(
         string $tableSpecId,
@@ -278,10 +280,6 @@ class Tables extends CompositeBase
 
     /**
      * Returns the distinct values for a table column, optionally filtered by query.
-     *
-     * @param array<string, mixed> $filter
-     *
-     * @return array<string, mixed[]>
      */
     public function distinct(string $tableSpecId, string $fieldName, array $filter = []): array
     {
@@ -417,6 +415,9 @@ class Tables extends CompositeBase
         $types = [];
         if (is_array($tableSpec[_ID_TYPE])) {
             foreach ($tableSpec[_ID_TYPE] as $type) {
+                if (!is_string($type)) {
+                    continue;
+                }
                 $types[] = ['rdf:type.u' => $this->labeller->qname_to_alias($type)];
                 $types[] = ['rdf:type.u' => $this->labeller->uri_to_alias($type)];
             }
@@ -519,8 +520,8 @@ class Tables extends CompositeBase
     /**
      * Count the number of documents in the spec that match $filters.
      *
-     * @param string               $tableSpec Table spec ID
-     * @param array<string, mixed> $filters   Query filters to get count on
+     * @param string $tableSpec Table spec ID
+     * @param array  $filters   Query filters to get count on
      *
      * @return int
      */
@@ -558,7 +559,7 @@ class Tables extends CompositeBase
 
         foreach ($specNames as $specName) {
             // Ignore any other types of specs that might have been passed in here
-            if (isset($specTypes[$specName])) {
+            if (is_string($specName) && isset($specTypes[$specName])) {
                 $this->config->getCollectionForTable($this->storeName, $specName)->deleteMany($query);
             }
         }
@@ -611,7 +612,7 @@ class Tables extends CompositeBase
      * If an exception in thrown because a field is too large to index, the field is
      * truncated and the save is retried.
      *
-     * @param array<string, mixed> $generatedRow the rows to save
+     * @param array $generatedRow the rows to save
      *
      * @throws \Exception
      */
@@ -640,7 +641,7 @@ class Tables extends CompositeBase
      * the total size of an index entry, which can include structural overhead
      * depending on the BSON type, must be less than 1024 bytes.
      *
-     * @param array<string, mixed> $generatedRow - Pass by reference so that the contents is truncated
+     * @param array $generatedRow - Pass by reference so that the contents is truncated
      */
     protected function truncateFields(Collection $collection, array &$generatedRow): void
     {
@@ -684,8 +685,8 @@ class Tables extends CompositeBase
     }
 
     /**
-     * @param array<string, mixed> $spec The table spec
-     * @param mixed[]              $dest The table row document to save
+     * @param array   $spec The table spec
+     * @param mixed[] $dest The table row document to save
      */
     protected function doComputedFields(array $spec, array &$dest): void
     {
@@ -704,9 +705,9 @@ class Tables extends CompositeBase
     }
 
     /**
-     * @param string               $function A defined computed value function
-     * @param array<string, mixed> $spec     The computed field spec
-     * @param array<string, mixed> $dest     The table row document to save
+     * @param string $function A defined computed value function
+     * @param array  $spec     The computed field spec
+     * @param array  $dest     The table row document to save
      *
      * @return mixed The computed value
      */
@@ -735,9 +736,6 @@ class Tables extends CompositeBase
     }
 
     /**
-     * @param array<int, mixed>    $equation
-     * @param array<string, mixed> $dest
-     *
      * @return float|int|null
      *
      * @throws \InvalidArgumentException
@@ -761,6 +759,9 @@ class Tables extends CompositeBase
         if (is_array($right)) {
             $right = $this->computeArithmeticValue($right, $dest);
         }
+
+        $left = is_numeric($left) ? $left + 0 : 0;
+        $right = is_numeric($right) ? $right + 0 : 0;
 
         switch ($equation[1]) {
             case '+':
@@ -796,8 +797,8 @@ class Tables extends CompositeBase
     }
 
     /**
-     * @param array<string, mixed> $replaceSpec The replace value spec
-     * @param array<string, mixed> $dest        The table row document to save
+     * @param array $replaceSpec The replace value spec
+     * @param array $dest        The table row document to save
      *
      * @return mixed
      */
@@ -818,12 +819,12 @@ class Tables extends CompositeBase
             $subject = $this->rewriteVariableValue($replaceSpec['subject'], $dest);
         }
 
-        return str_replace($search, $replace, $subject);
+        return str_replace($this->asStringOrArray($search), $this->asStringOrArray($replace), $this->asStringOrArray($subject));
     }
 
     /**
-     * @param array<string, mixed> $conditionalSpec The conditional spec
-     * @param array<string, mixed> $dest            The table row document to save
+     * @param array $conditionalSpec The conditional spec
+     * @param array $dest            The table row document to save
      *
      * @return mixed The computed value
      */
@@ -838,7 +839,7 @@ class Tables extends CompositeBase
                 $left = $this->rewriteVariableValue($conditionalSpec['if'][0], $dest);
             }
 
-            if (isset($conditionalSpec['if'][1])) {
+            if (isset($conditionalSpec['if'][1]) && is_string($conditionalSpec['if'][1])) {
                 $operator = $conditionalSpec['if'][1];
             }
 
@@ -869,9 +870,9 @@ class Tables extends CompositeBase
     }
 
     /**
-     * @param mixed                $value   The value to replace, if it contains a variable
-     * @param array<string, mixed> $dest    The table row document to save
-     * @param string|null          $setType Force the return to be set to specified type
+     * @param mixed       $value   The value to replace, if it contains a variable
+     * @param array       $dest    The table row document to save
+     * @param string|null $setType Force the return to be set to specified type
      *
      * @return mixed
      */
@@ -930,7 +931,9 @@ class Tables extends CompositeBase
 
         switch ($type) {
             case 'string':
-                $value = (string) $value;
+                if (is_scalar($value)) {
+                    $value = (string) $value;
+                }
 
                 break;
 
@@ -941,8 +944,12 @@ class Tables extends CompositeBase
                 break;
 
             case 'numeric':
-                if ((!is_int($value)) && !is_float($value)) {
-                    $value = $value == (string) (int) $value ? (int) $value : (float) $value;
+                if (!is_int($value) && !is_float($value)) {
+                    if (is_scalar($value)) {
+                        $value = $value == (string) (int) $value ? (int) $value : (float) $value;
+                    } else {
+                        $value = 0;
+                    }
                 }
 
                 break;
@@ -1003,7 +1010,13 @@ class Tables extends CompositeBase
 
             case 'contains':
             case 'not contains':
-                $bool = is_array($left) ? in_array($right, $left) : strpos((string) $left, (string) $right) !== false;
+                if (is_array($left)) {
+                    $bool = in_array($right, $left);
+                } else {
+                    $haystack = is_scalar($left) ? (string) $left : '';
+                    $needle = is_scalar($right) ? (string) $right : '';
+                    $bool = $needle !== '' && strpos($haystack, $needle) !== false;
+                }
 
                 $result = ($bool && $operator !== 'not contains');
 
@@ -1011,7 +1024,7 @@ class Tables extends CompositeBase
 
             case '~=':
             case '!~':
-                $match = preg_match($right, $left);
+                $match = is_string($right) && is_string($left) ? preg_match($right, $left) : 0;
                 $result = ($match > 0 && $operator !== '!~');
 
                 break;
@@ -1022,10 +1035,6 @@ class Tables extends CompositeBase
 
     /**
      * Add fields to a table row.
-     *
-     * @param array<string, mixed> $source
-     * @param array<string, mixed> $spec
-     * @param array<string, mixed> $dest
      */
     protected function addFields(array $source, array $spec, array &$dest): void
     {
@@ -1090,10 +1099,6 @@ class Tables extends CompositeBase
 
     /**
      * Generate values for a given predicate.
-     *
-     * @param array<string, mixed> $source
-     * @param array<string, mixed> $f
-     * @param array<string, mixed> $dest
      */
     protected function generateValues(array $source, array $f, string $predicate, array &$dest): void
     {
@@ -1307,6 +1312,22 @@ class Tables extends CompositeBase
     }
 
     /**
+     * Coerce a computed value into something str_replace() accepts.
+     *
+     * @param mixed $value
+     *
+     * @return array|string
+     */
+    private function asStringOrArray($value)
+    {
+        if (is_array($value) || is_string($value)) {
+            return $value;
+        }
+
+        return is_scalar($value) ? (string) $value : '';
+    }
+
+    /**
      * Apply a specific modifier
      * Options you can use are
      *      lowercase - no options
@@ -1316,7 +1337,7 @@ class Tables extends CompositeBase
      * @param string|string[]|\Stringable|\Stringable[] $value
      * @param array<string, bool|int|string>            $options
      *
-     * @return mixed
+     * @return mixed the modified value
      *
      * @throws \Exception
      */
@@ -1365,8 +1386,6 @@ class Tables extends CompositeBase
 
     /**
      * Apply a regex to the RDF property value defined in $value.
-     *
-     * @param array<string, mixed> $value
      *
      * @return false|int
      *
